Add additional checks to automatic image updates.

+++ install/chatgpt-proxy.sh
@@ -481,15 +481,19 @@ cat > /opt/script/go-chatgpt-api/AutoImageUp.sh << \EOF
 #!/usr/bin/env bash
 IMAGE_GOCHAT="linweiyuan/go-chatgpt-api"
 CURRENT_VERSION=$(docker image inspect $IMAGE_GOCHAT --format='{{index .RepoDigests 0}}' | grep -o 'sha256:[^"]*')
-LATEST_VERSION=$(curl -s "https://registry.hub.docker.com/v2/repositories/$IMAGE_GOCHAT/tags/latest" | jq -r '.digest')
+LATEST_VERSION=$(curl -s --max-time 60 "https://registry.hub.docker.com/v2/repositories/$IMAGE_GOCHAT/tags/latest" | jq -r '.digest')
 
-if [ "$CURRENT_VERSION" != "$LATEST_VERSION" ]; then
-  echo "go-chatgpt-api 镜像已更新，进行容器重启等操作..."
-  docker pull $IMAGE_GOCHAT
-  cd /data/go-chatgpt-api && docker-compose down && docker-compose up -d
-  docker rmi $(docker images -q --filter "dangling=true" --filter "reference=$IMAGE_GOCHAT") &>/dev/null
+if [[ -n $CURRENT_VERSION && -n $LATEST_VERSION ]]; then
+  if [ "$CURRENT_VERSION" != "$LATEST_VERSION" ]; then
+    echo "go-chatgpt-api 镜像已更新，进行容器重启等操作..."
+    docker pull "$IMAGE_GOCHAT"
+    cd /data/go-chatgpt-api && docker-compose down && docker-compose up -d
+    docker rmi $(docker images -q --filter "dangling=true" --filter "reference=$IMAGE_GOCHAT") &>/dev/null
+  else
+    echo "go-chatgpt-api 镜像无需更新"
+  fi
 else
-  echo "go-chatgpt-api 镜像无需更新"
+  echo "获取Images ID失败请稍后再试！"
 fi
 EOF
 chmod +x /opt/script/go-chatgpt-api/AutoImageUp.sh
